+++ include/libopencm3/stm32/common/flash_common_f234.h
@@ -74,7 +74,6 @@
 
 BEGIN_DECLS
 
-void flash_set_ws(uint32_t ws);
 void flash_unlock(void);
 void flash_lock(void);
 void flash_clear_pgperr_flag(void);
