+++ include/libopencm3/stm32/f0/flash.h
@@ -36,6 +36,7 @@
 #define LIBOPENCM3_FLASH_H
 /**@{*/
 
+#include <libopencm3/stm32/common/flash_common_all.h>
 #include <libopencm3/stm32/common/flash_common_f01.h>
 
 /* --- FLASH_OPTION values ------------------------------------------------- */
