+++ include/libopencm3/stm32/f1/flash.h
@@ -43,6 +43,7 @@
 
 /**@{*/
 
+#include <libopencm3/stm32/common/flash_common_all.h>
 #include <libopencm3/stm32/common/flash_common_f01.h>
 
 /* --- FLASH_OPTION bytes ------------------------------------------------- */
