+++ include/libopencm3/stm32/f2/flash.h
@@ -31,6 +31,7 @@
 #ifndef LIBOPENCM3_FLASH_H
 #define LIBOPENCM3_FLASH_H
 
+#include <libopencm3/stm32/common/flash_common_all.h>
 #include <libopencm3/stm32/common/flash_common_f24.h>
 
 #endif
