+++ lib/stm32/common/flash_common_all.c
@@ -1,14 +1,5 @@
-/** @defgroup flash_defines FLASH Defines
+/** @addtogroup flash_file
  *
- * @ingroup STM32F4xx_defines
- *
- * @brief Defined Constants and Types for the STM32F4xx FLASH Memory
- *
- * @version 1.0.0
- *
- * @date 14 January 2014
- *
- * LGPL License Terms @ref lgpl_license
  */
 
 /*
@@ -28,11 +19,19 @@
  * along with this library.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef LIBOPENCM3_FLASH_H
-#define LIBOPENCM3_FLASH_H
+/**@{*/
 
-#include <libopencm3/stm32/common/flash_common_all.h>
-#include <libopencm3/stm32/common/flash_common_f24.h>
+#include <libopencm3/stm32/flash.h>
 
-#endif
+void flash_prefetch_enable(void)
+{
+	FLASH_ACR |= FLASH_ACR_PRFTEN;
+}
 
+void flash_prefetch_disable(void)
+{
+	FLASH_ACR &= ~FLASH_ACR_PRFTEN;
+}
+
+
+/*@}*/
