+++ lib/stm32/f2/Makefile
@@ -43,8 +43,9 @@ OBJS            += crc_common_all.o dac_common_all.o dma_common_f24.o \
                    iwdg_common_all.o rtc_common_l1f024.o \
                    timer_common_all.o timer_common_f0234.o \
 		   timer_common_f24.o usart_common_all.o usart_common_f124.o \
-		   flash_common_f234.o flash_common_f24.o hash_common_f24.o \
+		   hash_common_f24.o \
 		   crypto_common_f24.o exti_common_all.o rcc_common_all.o
+OBJS		+= flash_common_all.o flash_common_f234.o flash_common_f24.o
 OBJS		+= rng_common_v1.o
 OBJS            += spi_common_all.o spi_common_v1.o spi_common_v1_frf.o
