+++ lib/stm32/f3/Makefile
@@ -42,9 +42,10 @@ OBJS		= rcc.o adc.o can.o pwr.o dma.o flash.o desig.o
 OBJS            += gpio_common_all.o gpio_common_f0234.o \
 		   dac_common_all.o crc_common_all.o crc_v2.o \
 		   iwdg_common_all.o pwr_common_v1.o dma_common_l1f013.o\
-		   timer_common_all.o timer_common_f0234.o flash_common_f234.o \
-		   flash.o exti_common_all.o rcc_common_all.o
+		   timer_common_all.o timer_common_f0234.o \
+		   exti_common_all.o rcc_common_all.o
 OBJS		+= adc_common_v2.o adc_common_v2_multi.o
+OBJS		+= flash_common_all.o flash_common_f234.o
 OBJS		+= usart_common_v2.o usart_common_all.o
 OBJS		+= i2c_common_v2.o
 OBJS		+= spi_common_all.o spi_common_v2.o
