+++ lib/stm32/f3/flash.c
@@ -58,25 +58,5 @@ void flash_clear_status_flags(void)
 	flash_clear_eop_flag();
 }
 
-/**
- * Enable the FLASH Prefetch Buffer
- * This buffer is used for instruction fetches and is enabled by default after
- * reset.
- */
-void flash_prefetch_enable(void)
-{
-	FLASH_ACR |= FLASH_ACR_PRFTBE;
-}
-
-/**
- * Disable the FLASH Prefetch Buffer
- * This buffer is used for instruction fetches and is enabled by default after
- * reset.
- */
-void flash_prefetch_disable(void)
-{
-	FLASH_ACR &= ~FLASH_ACR_PRFTBE;
-}
-
 /**@}*/
