+++ lib/stm32/l0/Makefile
@@ -46,7 +46,7 @@ OBJS		+= adc_common_v2.o
 OBJS		+= crs_common_all.o
 OBJS		+= dma_common_l1f013.o
 OBJS		+= exti_common_all.o
-OBJS		+= flash.o flash_common_l01.o
+OBJS		+= flash.o flash_common_all.o flash_common_l01.o
 OBJS		+= i2c_common_v2.o
 OBJS		+= rng_common_v1.o
 OBJS		+= usart_common_all.o usart_common_v2.o
